backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and Security
pyjwt==2.8.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Payment Processing
stripe==7.8.0

# Configuration and Environment
pydantic==1.10.13
pydantic[dotenv]==1.10.13
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Validation and Serialization
email-validator==2.1.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging
loguru==0.7.2

# UUID
uuid==1.30

# Random and String utilities (built-in, but listed for clarity)
# string - built-in
# random - built-in
# datetime - built-in
# typing - built-in

# Optional: For production deployment
gunicorn==21.2.0

# Optional: For monitoring and health checks
prometheus-client==0.19.0